Exclude vendor sub-dir from shell linting

+++ .buildkite/steps/lint.sh
@@ -4,6 +4,7 @@ set -euo pipefail
 grep -rl '^#!/.*sh' . | while read -r file ; do
   [[ $file =~ \.git ]] && continue
   [[ $file =~ init\.d ]] && continue
+  [[ $file =~ vendor ]] && continue
 
   echo "Processing $file"
   docker run --rm -v "$PWD:/mnt" koalaman/shellcheck "$file"
